Modernised TableView example project and made compatible with the latest version of RN
