+++ examples/TableViewDemo/ios/TableViewDemo/Images.xcassets/AppIcon.appiconset/Contents.json
