+++ examples/TableViewDemo/package.json
@@ -1,17 +1,28 @@
 {
   "name": "TableViewDemo",
-  "version": "0.0.1",
+  "version": "0.1.0",
   "private": true,
+  "devDependencies": {
+    "babel-preset-react-native-stage-0": "^1.0.1",
+    "jest-expo": "^21.0.2",
+    "react-test-renderer": "16.0.0-alpha.12"
+  },
   "scripts": {
-    "start": "node_modules/react-native/packager/packager.sh"
+    "start": "react-native start",
+    "android": "react-native run-android",
+    "ios": "react-native run-ios",
+    "test": "node node_modules/jest/bin/jest.js --watch"
+  },
+  "jest": {
+    "preset": "jest-expo"
   },
   "dependencies": {
-    "firebase": "^2.4.0",
-    "react-native": "^0.24.0",
-    "react": "0.14.5",
-    "react-native-navbar": "^1.2.0",
-    "react-native-router-flux": "^2.1.8",
-    "react-native-tableview": "1.5.0",
-    "react-native-tabs": "^0.1.10"
+    "firebase": "^4.5.0",
+    "react": "^16.0.0",
+    "react-native": "^0.49.3",
+    "react-native-navbar": "^2.0.1",
+    "react-native-router-flux": "^4.0.0-beta.22",
+    "react-native-tableview": "https://github.com/houserater/react-native-tableview.git#hank.brekke/react-16-upgrade",
+    "react-native-tabs": "^1.0.9"
   }
 }
